Add camera line tracing and calulating error for the PID.
Added Camera, PID, System etc structs and enable the camera to track the
black line and calculate error by taking a line of pixels in the centre
of the screen, determining if they are black or white, then averaging
their positions. In the PID we subtract the position of the line from
the horizontal centre of the screen to yield an error value for which we
can start to correct.